requirements.txt
# Core ML & NLP
torch==2.6.0
sentence-transformers==5.2.3
langchain==1.2.10
langchain-ollama==1.0.1
faiss-cpu==1.13.2

# Hugging Face
datasets==2.18.0
dill==0.3.8
fsspec==2024.2.0
pyarrow==20.0.0
multiprocess==0.70.15
xxhash==3.5.0

# Data Science & Utilities
numpy==2.2.4     # MUST use 2.x for Python 3.13
tqdm==4.66.0
pandas==2.2.3

# Network
requests==2.32.5
aiohttp==3.12.0
pyyaml==6.0.2